requirements.txt
diffusers==0.32.1
transformers==4.48.0
opencv-python
safetensors
pillow==10.3.0
einops
peft
imageio
omegaconf
ultralytics
tqdm==4.67.1
protobuf==3.20.2
torch==2.3.0
torchvision==0.18.0
moviepy==1.0.3
ftfy